Fix a few more misleadingly named variables
BW-5687
http://makerbot.atlassian.net/browse/BW-5687

MaterialName is for display names and Material is for API names.  We
still had some misleadingly named variables which caused the use of
an API name in a display context to go unnoticed.  This commit adds
the transform function to convert these API names into display names,
and also renames the variables to reflect that they contain API names.