Create UI digitizer and lcd panel inspection tool (#489)
BW-5683
http://makerbot.atlassian.net/browse/BW-5683

Test is added to the Advanced Settings Page and displays 8 white circles, and keeps track of user tracked touches for each circle. The test will display the average offset in pixels for X and Y dependent on where users touch panel. It additionally shows the offset magnitude, which is what the technician will use.