Fix blank screen in print page on clicking change material
SwipeViews used as page containers are implemented using the
LoggingSwipeView component which has its own SwipeToItem function,
which in addition to swiping does a few more things viz. logging
actions and toggling visibilities of the pages.

The bug here was that the swipe was being perforemed directly
by manipulating the current index, bypassing the swipeToItem
function, which was causing the visibility issue.

BW-5681
https://makerbot.atlassian.net/browse/BW-5681