Ensure survey is asked after atleast 3 months of usage
For a brand new printer or a printer just reset to factory the
survey is asked only after 3 months of using the printer. The
3 month period is counted from the last time the FRE has been
interacted with. For existing users updating their firmware
this survey will be prompted at the end of the first print.

This commit also switches from saving the survey submission
time to saving the next survey due time for better readability
and representation of the use case of making the survey due 3
months out from first use.

BW-5584
https://makerbot.atlassian.net/browse/BW-5584