Allow load for sunflower when filament present but null material
BW-5700
http://makerbot.atlassian.net/browse/BW-5700

For Sunflower, if there is no material information in loaded filaments, but the filament switch is triggered we want the user to be able to load and select materials (via the load button). The unload button remains enabled if the filament switch is triggered regardless of whether we display load or purge.