Use supported materials from kaiten for materials check
BW-5680
http://makerbot.atlassian.net/browse/BW-5680

Replaced the "goodMaterialsList" with the supported materials list that is created in kaiten and used for the sunflower load process. Kaiten will handle the check for high temp materials on fire, to prevent high temp materials from being in the supported materials list. Updated places where this is used.