Added cancelling state for zip logs process
BW-5602
http://makerbot.atlassian.net/browse/BW-5602

Noticed that when cancelling, the process would have to wait for the "write" method to be completed before cancelling the process and removing the files safely. This causes issues (from a user standpoint) because the popup will close but the process will still be cancelling. To remove any confusion an additional page was added to the popup so the UI will prompt the user to wait while the process is cancelling before removing the USB.
Also added the changes proposed during code review.