Use labs workflow for mid-print loading for sunflower as well
Ask the user to load the current print material in the material page.
When acknowledging the OOF error. Skip the filament bay handling screens
and jump straight to the material page after clearing the error (same
behavior as labs extruder)

This is also similar to what's displayed on Method/X on OOF but the text
there is in the filament color field which isn't shown at all in sunflower
so instead this is shown in the material name field.

BW-5716
https://makerbot.atlassian.net/browse/BW-5716