Allow skipping feedback in all cases (print successful/cancelled/failure)
This is to allow users to be able to end a print in case the feedback
buttons don't show up due to a previously observed UI bug. This commit
also has fixes to address this missing buttons bug, but the print can
now be ended even if the buttons go missing.

BW-5676
https://makerbot.atlassian.net/browse/BW-5676