Wifi UI Enahncements and Custom Popup Use Changes
BW-5606
http://makerbot.atlassian.net/browse/BW-5606

Wifi UI Enhancements following the spec. Changes to the Menu Button component to use for the Wifi items, and removed the WifiButton because it isn't being used anymore.

Custom Popup Changes to define size of the popup by the contents in the popup, does not change the functionality of other popups already created. Changed some new custom popups using this model to size it in this other way. Also added the popup images inside of the component so that they don't have to be changed in multiple places if they're using the popup images.