Loading flow fixes for labs
BW-5747
http://makerbot.atlassian.net/browse/BW-5747

A lot of our labs specific behavior during loading was gated behind a
variable called isExternalLoadUnload.  The concept of being an external
load is separate from being a labs extruder and we now have external
loading on sunflower without labs, so this variable was renamed.  More
importantly though, this variable wasn't actually being assigned to line
up with either of these concepts.  Instead we have been assigning this
based on whether we provide a custom temperature override to kaiten,
which was equivalent to using a labs extruder in 1.11 but is now usually
not the case for labs extruders.  So now we just bind this directly to
whether we have a labs extruder installed.

As part of this update I found that the logic for the retry button was
also not updated to handle the changes in how custom temperature was
handled.  So this logic was updated and dramatically simplified.

This just brings the behavior for labs back to what it was in 1.11.
There are clearly still a lot of changes that have to happen here for
sunflower support and probably also some opportunities to improve the
flow for labs.