Redesign time page; Updates to set things up for the new sizes
It was really annoying to design this screen for the new size without
the entire UI (all components/pages) actually being able to accomodate
the new sizing, so I changed quite a bit of things throughout to make
everything fit into the new layout, but there might be a few things
that dont look right due to being pushed 12px below and also overlayed
by a pretty large fade area below the already big topbar.

Most of the chnges here are to make the entire UI accomodate the
new size which is 800x408 (old: 800x420) for the content area and
800x72 (old: 800x60) for the top bar. There is also a 30px high fade
area below the topbar. To get things going I just resized the existing
topbar to the new size and also added the fade below.

So the good thing is all pages made from now on can be made for the
new size without any guesswork and everything will show up correctly
without any offsets.

I also added the dimensions to the settings json since a lot of them
seem to be slighly more standardized in the new figma design, but since
the UI program needs to be run to actually read the dimensions it was
inconvenient to not be able to see the pages in the correct dimensions
in the design view in qt creator and I'm sure there is a better way to
visualize responsive designs in qt creator, so I just commented out
getting the dimensions from settings for now but there is an opportunity
now to do this as part of the redesign work.

BW-5599
https://makerbot.atlassian.net/browse/BW-5599