Materials screen labs fixes
BW-5747
http://makerbot.atlassian.net/browse/BW-5747

When the labs extruder is loaded externally but does not know what type
of filament is loaded (and there is no filament in the bay), the
materials screen will now say only LABS MATERIAL LOADED, and will use
the same icon as we use in all other cases where filament is externally
loaded.

We do this by continuing to have "loaded filament" report "unknown" for
both unloaded filament and any case where kaiten does not know filament
type, and then try to use this field as little as possible, since it is
not really consistent between method and sunflower.

I also simplified the color/filament amount logic, because we don't need
to have UNKNOWN MATERIAL under the actual material name all the time on
sunflower and because having two lines containing empty strings under
LABS MATERIAL LOADED breaks the layout a bit.