Scroll the ethernet text with the wifi list
The idea here is that the reset button is always on top of the ethernet
text and when we are showing a wifi list the ethernet text is the header
element of the list (so that it scrolls with the list).  We make two
copies of the ethernet text so that we also have a non scrolling version
to show when the list is not visible (though I am not entirely clear on
why we can't always show the list or whether we really need the ethernet
text in the cases where we don't show the list).

This also changes around the logic to disable attempted use of wifi while
ethernet is attached because this really does not work well, and changes
the scanning logic to separately track whether we are actively scanning
and whether we are currently connected, because apparently the rescan
button has been completely unusable since 1.5 if you aren't already
connected to wifi.  This also fixes the top bar title.