Compile qml at build time instead of runtime
QML modules are compiled before use.  Currently this just happens every
time that we load the UI application (there is a mechanism for caching
these files but that does not work due to the read only filesystem).
However there is support for compiling these modules at compile time
and bundling the compiled versions into the binary along with the QML
files themselves (which according to the docs are also used in some
cases).

This commit adds compiled qml into the binary at compile time so that
the compilation step can be skipped at runtime.  This is accomplished
through a QML module (that we already build with buildroot) which just
replaces the qt5_add_resources call with a call that also compiles qml
for every qml file listed in a qrc file.

I measured one time each the time to restart the UI when the system is
idle before and after this change and got 46 seconds before and 33
seconds after.  The file size was 32MB before and 34MB after.  The time
improvement isn't actually that impressive and with only one sample
isn't even statisticaly significant.

But I still think that this is worth doing because we get a build time
check that all of our qml files are valid.