Keep track of door and lid errors disabled separately
Moving the build plate is now a kaiten process which can error
out. This fixes a case where the door open warning popup was not
shown on the UI because the lid error was disabled and the process
was silently erroring out due to the door being open without any
indication to the user. Other toolhead errors will also silently
fail the move build plate process and we might have to either
ignore those errors or communicate better to the user.