Remove dimensions from settings
Qt doesn't seem to be picking up dimensions defined in parent qml
files when in the designer view, so it wasn't really adding any value
to what we already have now. Instead to unify this setting, I just put
the content  dimensions (800x408) in the LoggingItem which is the base
item for all pages so they dont have to be defined in every page.Bbut to
still see the dimensions in the design view you would have to hard-code
it in the specific qml file during development and hopefully remember
to remove it before committing.