New calibration UI
Also add AnimatedImage to ContentLeftSide template which can be used
to put gifs on the left side of the screen.

One weird bug I noticed was the calibration page had trouble resetting
back to the base state at the end of the process even when explicitly
assigned to. It would go the base state, but the property changes were
only partially applied. The solution was to clone the base state with
the default property changes and that seemed to work. We rely on the
built-in base state all over the UI for every process and everywhere
else states are used and this was the first time I encounterd this
bug. I am not sure if it will be a recurring thing across those too
but atleast we have a workaround. The partial property changes were
only applied on the newly added template components so I have no
idea whats happening.

BW-5768
https://makerbot.atlassian.net/browse/BW-5768