Complete the blocking UI flow
Hook up network change signals and firmware available signals so that
the network update flow actualy flows.  The general flow is that as
soon as kaiten indicates that firmware is available, the "connect to
network" button directs straight to the firmware update screens, and
otherwise if we have an internet connection we ask kaiten to check for
firmware and give a little popup to indicate that we are checking for
firmware.  Only in the case where we have no internet yet will this
button go to directly the wifi screen, though failing to find firmware
when on wifi will also direct users to the wifi screen so they can fix
any wifi issues they are having.