Change "CONNECT TO NETWORK" text
This button now more explicitly says what it does -- start downloading
and installing a firmware update if this is possible or scan for wifi
networks if there is no internet connection.  In general we should not
be turning on the wifi radio if the user hasn't explicitly said that
they want to use wifi, so changing the text of this button to be a wifi
button whenever pressing this button would enable wifi seems like the
most straightforward way to address this.