Later now means 3 months later
BW-5766
http://makerbot.atlassian.net/browse/BW-5766

I am assuming that the original intent of the "Later" button was just
"ask me again on the next print" but really who is pressing this button
who actually wants to get this popup again on the next print?  Maybe
we can do some future release where we have some more nuanced behavior
for this button, but for now we just want a fast and low risk way to
make sure we aren't spamming this popup to users who don't want to see
it.