Reset pages to base page without affecting the current item
The code to "reset" the numerous nested swipe views just swipes each of
them to their respective base pages -- a side effect of which that we end
up setting the current item to the target page during each of those swipes
which is really undesirable. To prevent this, I added a flag to perform a
swipe without setting the destination as the current item.
I also added resetting the mainswipeview to the reset pages function to
get the UI into a baseline state after which we can unambigously move the
UI to the desired page.