Log clicks with the released signal emission
Connecting a logging function to the clicked signal seems to not be
ideal as the connected function is only called after the signal
handlers code has been executed which isnt ideal when looking at the
logs as we first want the click to be logged followed by the events
caused by the click and not the other way around. Logging using the
released signal seems to capture as much as the clicked signal but also
in the correct order. Both of these arent logging when a button is
pressed but the finger is moved out of the button area and then released
so there is no loss of functionality switching to the released signal
to handle logging.