fix main menu missing print progress icon
BW-5820
http://makerbot.atlassian.net/browse/BW-5820

An uncaught merge error resulting from the reordering of code blocks allowed an error into the
main menu: the print progress icon in the main menu went "missing" while printing.

The problem was the interaction between the templating system for each of the 3-across icons, only
the Print Icon ever being switched over to a different icon (while printing), and the recently added
auto-flowing ColumnLayout to the icon template.

In the name of expediency, a solution for a template-able switch-able icon compatible with ColumnLayout
auto-reflows will be left for future updates, and this pull request will go back to the era of magic
numbers for hitting the design spec appearance.