Improve unsupported extruder/extruder-combo warning
This commit adds the abilty to keep track of a list of supported extruders
for each slot as reported by kaiten similar to the supported materials list.
When an incompatible extruder is installed the extruder mismatch popup
now shows the list of supported extruders dynamically for that printer while
also considering the extruder in the other slot. So supported extruder for a
slot also depends on the extruder present in the other slot. This is required
now as the Method XL at launch has some restrictions on what extruders can be
used. The only restriction until now on Method/X was to not allow normal and
hot extruders combo.

The extruder combo validity logic has been removed from the UI and is now
evaluated on kaiten side so the new restrictions for Method XL and the
existing restrictions for Method/X should all just work.

The extruder compatibilty for a single slot is determined by looking at the
tool_type_correct flag sent in the system info notification.
The extruder combo compatibility is determined by looking at the
can_pair_tools field sent in the extruder change notification as well as
the extruders config update notification which are also used for building the
the supported extruders and supported materials lists.

BW-5791
https://makerbot.atlassian.net/browse/BW-5791