Skip precompiling QML assets when building on Qt creator
Precompiling QML assets is pretty convenient for toolchain release builds
to identify broken code before actually running the program on the printer
and waiting for runtime compilation to fail.

For Qt creator builds this however seems to introduce some quirks like the
stale cache issue that prevents building without reconfiguring cmake
(I think), clearing previous build dirs (which is generally a good thing to do
anyway) after switching to a branch with different qml assets than what the
old compilation cache contains. This cache issue is also present for
toolchian release builds now. Another issue with precompiling is the inability
to click through the error messages from the output panel in Qt creator and be
taken to the offending file and line number which is pretty convenient. With
precompiling, Qt creator tries to access the compiled object file and fails.
Another minor issue I noticed was it also slightly broke the standard find
feature to find strings across all files in the current project in Qt creator
though there was an easy workaround for it.

Precompiling doesn't seem to provide a massive advantage for Qt Creator builds
as runtime compilation is pretty quick on desktop processors with the errors
printed in the output panel immediately.