Add an FRE step for sunflower unpacking
BW-5846
http://ultimaker.atlassian.net/browse/BW-5846

We need a step guiding users through removing one box, then lifting the
build plate so they can remove a second box.  The spec also calls for a
popup to try to make sure that the user was really paying attention when
we asked them to remove the packaging, and another popup to make sure
that they closed the door, plus a screen to remind them to close the
door if they open it before the build plate moves up.  We just move the
build plate up until the top endstop is triggered.

The styling on the popup was just copied from the "printer is busy"
popup since that popup looks pretty much identical -- I have no idea how
necessary any of the layout commands there actually are.

This is missing any checking for a failure to start the process to move
the build plate because this really shouldn't fail to start, and we just
assume any error that cancels the process is a door error since nothing
else should error this out.  We also don't validate that we actually hit
the endstop, so any user that doesn't remove the top box is going to
get stuck at a the screen asking them to remove the bottom box.  But it
isn't really clear what we would want the user to do in this case...

I don't really get the pattern of using "base state" for one of the
states in every stateful component instead of a properly descriptive
name for every single state.  This state isn't applied by default so
you still have to manually make sure that the component is in this
state, and you also can't compare against "base state" to see if no
other state has been assigned yet, so I don't really see any way in
which qt treats this string as special.  I just used descriptive names
for every state and explicitly specified the initial state in QML and
everything seems to work fine.

I also used property changes to explicitly tie popup visibility to the
state of the top level component, and I don't see any issues from doing
this instead of using open() and close() on the popup.

I didn't really like the setup where for an FRE step which we might
skip, the "next step" callout from the step before needs to check if
we need to skip, and if so explicitly move to the step after next.
So I added a mechanism to the FRE tracker to tell it to when to skip
over steps when we request that it go to the next step.  (I did not
change any existing logic to use this mechanism though since that
would be a bit outside the scope of this ticket.)