Updates for name printer and cloud connect
BW-5845
http://ultimaker.atlassian.net/browse/BW-5845

The name printer screen now has progressing to the next step as the
default option and actually changing the machine name as the secondary
option.  It also now occurs after the test print as part of "connect".

For the cloud connect step I scrapped the custom separate FRE only
component and we use the same component that we use during normal
operation.  We also seamlessly skip over this step if we are not
connected to the internet.

The logic for skipping over the network connection and firmware update
steps if they were not connected was fairly janky and directly
targetted the name printer step.  So I took this opportunity to make
these steps also get skipped according to the new step skipping
mechanism.