FRE updates to assisted leveling
BW-5841
http://ultimaker.atlassian.net/browse/BW-5841

Add a help popup for this step (which is the same generic fre help popup
used throughout the fre) and add a separate intro screen that is just
for the FRE which is a little bit clearer about which hex keys to use.

Updating the non-FRE specific behavior is a bit outside the scope of the
ticket, but I did go ahead and update the text on the remove build plate
and locate hex screws screens.  Replacing the image on that screen with
a gif (but only making it a gif on sunflower) is much more of a pain and
kind of breaks our getImageForPrinter architecture so I did not add that
change to the scope of this ticket.