Add a new material case setup step for sunflower
BW-5843
http://ultimaker.atlassian.net/browse/BW-5843

This adds a new FRE step for setting up the material case on sunflower.
We use the same mechanism as used for the packing removal step to ensure
that this only shows up in the FRE for sunflower.  The step is its own
component, and it lives in the extruder settings swipe view under the
assumption that we will add an entry there to let us access this flow
outside of the FRE.

The flow is just a sequence of screens describing how to set up the
material case, so we use the altBack functionality to let the back
button step backward through this sequence and back to the FrePage
entry for this flow.

Every screen is this flow is supposed to have a help link to the main
FRE help flow.  I added the link to the "fre" state of the help popup
here, but the main FRE help flow is implemented in a different PR and I
don't really feel like making a copy of it here -- this should work just
fine once both PRs are merged.