also updates non-magma wait-to-cool warn screen
BW-5836
http://ultimaker.atlassian.net/browse/BW-5836

Also applied the new content(Left/Right)Side templates to the non-sunflower wait-to-cool screens.

Separated the 2 wait-to-cool timers to operate on different total times, have different exit conditions, etc.

Added 2 quality-of-life changes to logs, one is tagging of contentRightSide button titles to appear more like other logged button titles, and the other was to remove 2 anchor settings that were spamming the logs.