Add skips to FRE + step skipping bugfix
BW-5854
http://ultimaker.atlassian.net/browse/BW-5854

This adds a skip button to every screen that doesn't have one.  I tried
to make showing the skip button the default for every state and then
only states that explicitly didn't want to show this button would
override this, but for some reason every state needs to be explicit
about this component being visible for it to show up.  The skip popup
should really only ever need to change the step if we are invoking it
from the FRE intro screen, so the callout to step specific cleanup
was made conditional on having actually progressed past the intro
screen.

I also noticed a bug while testing this that we are only updating the
set of enabled FRE steps based on bot model properties when those
properties change, and if our firmware is up to date or if we boot up
offline and stay offline, these properties never actually change (in
contrast, the bot type property is initialized to "unknown" so it
should always fire off a change event).  So I added some code to make
sure that we set FRE step enablement based on initial values once we
are finished setting up the change event listeners.