Allow a button to be disabled without affecting the adjacent help button
We just add a new style to the button that effectively mimics disabling the
button visually without affecting the help button beside. The one catch is
that we need to explicity handle suppressing this buttons onClicked action
when it is "disabled" using this style but it is pretty easy to do this
locally for the specific context where we want this behavior of the button
being disabled and the help button enabled.