Modify textHeader1 to exclusively display waiting status + simplifications
This waiting status is also used in the material loading page so I wanted
to get this in a good shape on develop before merging develop into the
material loading branch. textHeader1 was only being used in the ErrorScreen
at one place and we can use the textHeader to display the textHeader1 contents
too by just using a newline character.

I also went ahead and did a lot of cleanups notably reusing the extruder API
name to marketing name translation function we already have for the attach
extruder flow as well.