Move metadata json manipulations out of the asynchronous callback function
The code we have to convert the chamber temperature to build plane
temperature from the meta.json seems potentially intensive to be run in
the network request thread completion callback function for every file
in the cloud queue list since a lot of these threads finish in parallel
at the same time. It also seems unnecessary to do this at this point as
the buildplane temperature isnt displayed in the list view. It was earlier
displayed in the file info page which can only be accessed by clicking
through the file, but even that seems to have been removed after the
latest redesign so we will just defer doing this until the user actually
clicks on a file.