Make temperature status component more generic
The temperature status component was designed with the extruders status
at first in mind and thus tightly tied to it but unfortunately when it
was used for other intended use cases people just didnt bother updating
its template and using it effectively.

Also cleanup its usage in the print status page and the wait to cool
screen.