Rework first page of PrintStatusView; add externalprint process downloading step
This page was honestly a mess with weird conditions and unnecessary elements
some being reused to display completely unrelated information from what their
original intent was. I reworked this to use a standard format and make it less
confusing. Also using auto layouts now for this screen. I also spent a good
amount of testing this page in different print steps to make sure nothing
looked weird.

While testing starting a cloud queue print both directly from the printer
and from cloudprint I noticed a new kaiten process step in the extrenal
print process that hadnt been added to a UI state, so added it. We have
three steps for file transfers now that we should make a new stateType on
the UI for it and not put it into the Loading stateType which will make
some of the logic on this page cleaner.