Make the printing drawer not overlap with the top bar text
I dont know what's the deal with the drawers we have for printing and
material loading/unloading. They havent been touched in a while and
I heard we are removing them with the redesign but the functionality
is all there. In any case this commit just makes the drawer not
overlap with the taller top bar when pulled down.