Add a ChamberBuildPlane component to TemperatureStatus
We use the chamber build plane temperature when talking about the chamber
temperature during a print. Easiest way seems to be to add that as a separate
virtual component to the TemperatureStatus. I think the print process is the
only place where we communicate the build plane temperatures, the rest of the
places drying cycle, anneal print all use the regular chamber temperature, but
those screens are yet to be updated to new designs with the TemperatureStatus
component.