Please dont run an infinite timer
I was profiling the UI on Qt creator and stumbled upon this infinite timer
running at 100ms interval. It took about 120us for this to be serviced on
my laptop. On the printer, as suspected, this took anywhere from a few ms
to few tens of ms across runs. QML Timers are synchronized with the screen
render thread which updates approximately every 120ms or so, so that is
technically the minimum interval that can be acheived with QML timers.

More importantly the feature that this is being used for-- to show the time
on the top bar, isnt really to put a high precision clock on the UI, rather
to make it easier to correlate events across videos and logs provided to
customer support so I think we can manage comfortably with a 1 minute error
margin.