Put back filament bay OOF error resolution instructions
In the filament bay OOF error screen, as part of the redesign, only
the text on the right container was changed but the behavior of the
button on that page wasnt changed as there was no instructions or
workflow provided in figma for it. Specifically I think the design
team wanted the LOAD MATERIAL button on that page to just take us
to the material page, but the existing button logic is to only enable
that button when a matching material spool as the current print material
is placed on the bay and upon clicking that button, it directly starts
the material loading activity from this screen. The requirement to put
the matching material was earlier being communicated to the user but has
now been removed so this button is disabled with no indication when the
spool is removed from the bay which is confusing.

This button should really be a dumb button that takes us to the material
page but with the current logic it should definitely keep the instructions
that were removed. This is what happens when the design team has no clue
of the existing system behavior before coming up with design changes.