Do not show labs extruder loading instructions on Magma
This became slighlty more invloved that I initially thought with the
addition of a new waiting for filament step in kaiten. Earlier upon
starting a load process we unconditionally went to the preheating
step and started heating the extruders while waiting for the user to
insert filament, so this screen vivsibility condition was tied to the
preheating step as well as temperature checks likley due to some some
edge cases.

This commit updates this screens visibility logic to account for the new
kaiten and handle cases like not showing this screen when purging for
a good UX. Of course also not showing the screen on Method XL.

BW-5856
https://ultimaker.atlassian.net/browse/BW-5856