Lower the build plate again after unpacking
BW-5865
http://ultimaker.atlassian.net/browse/BW-5865

We don't want to just leave the build plate at the top of the chamber
because it makes it hard to remove the build plate when asked to by the
assisted leveling process.  Dropping the build plate 100mm gives plenty
of room to get the build plate off but also takes a lot less time than
dropping to the bottom.  And a user can open the door a few times and
resume without hitting the bottom (we don't currently halt at the bottom
endstop or track how far we dropped before the door opened).

We implement this flow by leaving all of the current states in place
(with some renamings) and adding a separate state variable to
differentiate between raising and lowering the build plate.