Restyled and hid the RETRY TEST PRINT button
BW-5896
http://ultimaker.atlassian.net/browse/BW-5896

I spoke with Praveen and instead of removing the button entirely, we are just going to hide it incase we would like to add back the functionality. As part of this I restyled the button for the completed and failed pages in the print status view. But currently this will not be visible.