Clean up the error screen page;Fix chamber fan failure error screen
I originally wanted to just fix the chamber fan failure error screen that
was broken, but ended up simplifying the whole page.

The main update is that we make the base state a blank canvas and then
just specify the relevant elements in the different error screens. Since
the states are transformations applied on the base state this approach is
clean and intuitive without any guesswork of what the base state might be.
This works nicely for the ErrorScreen component as the base state isnt used
at all and every error is just mapped to a state on this page unlike most
other pages on the UI that currently use the base state to refer to the
landing screen for that component.

I also spent a good amount of time testing these for visibility issues.

BW-5904
https://ultimaker.atlassian.net/browse/BW-5904