Fix resume print button prematurely enabled on the door/lid error screens
This was an issue Mignatti reported a while back. The states for both these
error screens were overriding the enabled condition for the resume button
that caused this bug. The enabled condition for this button is globally
defined in the .qml file i.e. not until kaiten gets to the paused step.

BW-5904
https://ultimaker.atlassian.net/browse/BW-5904