Add final QR codes
BW-5857
http://ultimaker.atlassian.net/browse/BW-5857

This sets QR codes for all help screens, for the FRE intro screen, and
for the manual firmware download screen.  (There was also a Figma
component for a generic QR code linking to support, but I couldn't
find any figma flows that actually used that components.)

All of the help screens aside from two were XL specific and could
therefore reasonably just link to the XL setup flow.  One help screen
was specifically for material compatibility which we have a specific
link for.  But the instructions for cutting the filament were written
to include a help link for all machine types, and the only link we have
that would address that is the XL setup guide.  So I removed the help
icon for all non-XL printers and made the qr code also link to XL
setup.link

All of the codes were checked in at 230 by 230 pixels, which is the
size that the codes are supposed to be in content left side.  Putting
the same size image in the help popup automatically scales it down to
a very similar size to what is in figma, and it doesn't seem to have
any scaling issues, so I stuck with just the one image size.

The firmware update page also needed some updates to not position the
QR code super awkwardly.  I switched the image over to use content left
side to get a good position, but I did not switch over the rest of the
update flow to use this component.